requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0

# AI/LLM integration
openai==1.6.1
google-generativeai==0.3.2

# Report generation
python-pptx==0.6.23
reportlab==4.0.7
Pillow==10.1.0

# Optional: faster data processing
# polars==0.20.2

# Utilities
python-dotenv==1.0.0
openpyxl==3.1.2
